+++ crates/api/api/src/post/mod_update.rs
@@ -6,12 +6,7 @@ use lemmy_api_utils::{
   context::LemmyContext,
   plugins::{plugin_hook_after, plugin_hook_before},
   send_activity::{ActivityChannel, SendActivityData},
-  utils::{
-    check_community_user_action,
-    check_is_mod_or_admin,
-    check_nsfw_allowed,
-    update_post_tags,
-  },
+  utils::{check_community_user_action, check_nsfw_allowed, is_mod_or_admin, update_post_tags},
 };
 use lemmy_db_schema::source::post::{Post, PostUpdateForm};
 use lemmy_db_views_local_user::LocalUserView;
@@ -46,7 +41,7 @@ pub async fn mod_edit_post(
   let community = orig_post.community;
 
   check_community_user_action(&local_user_view, &community, &mut context.pool()).await?;
-  check_is_mod_or_admin(&mut context.pool(), local_user_view.person.id, community.id).await?;
+  is_mod_or_admin(&mut context.pool(), &local_user_view, community.id).await?;
 
   let mut post_form = PostUpdateForm {
     nsfw: data.nsfw,
