+++ crates/api/api_utils/src/utils.rs
@@ -75,51 +75,25 @@ use webmention::{Webmention, WebmentionError};
 
 pub const AUTH_COOKIE_NAME: &str = "jwt";
 
-pub async fn check_is_mod_or_admin(
-  pool: &mut DbPool<'_>,
-  person_id: PersonId,
-  community_id: CommunityId,
-) -> LemmyResult<()> {
-  let is_mod = CommunityModeratorView::check_is_community_moderator(pool, community_id, person_id)
-    .await
-    .is_ok();
-  let is_admin = LocalUserView::read_person(pool, person_id)
-    .await
-    .is_ok_and(|t| t.local_user.admin);
-
-  if is_mod || is_admin {
-    Ok(())
-  } else {
-    Err(LemmyErrorType::NotAModOrAdmin.into())
-  }
-}
-
-/// Checks if a person is an admin, or moderator of any community.
-pub(crate) async fn check_is_mod_of_any_or_admin(
-  pool: &mut DbPool<'_>,
-  person_id: PersonId,
-) -> LemmyResult<()> {
-  let is_mod_of_any = CommunityModeratorView::is_community_moderator_of_any(pool, person_id)
-    .await
-    .is_ok();
-  let is_admin = LocalUserView::read_person(pool, person_id)
-    .await
-    .is_ok_and(|t| t.local_user.admin);
-
-  if is_mod_of_any || is_admin {
-    Ok(())
-  } else {
-    Err(LemmyErrorType::NotAModOrAdmin.into())
-  }
-}
-
 pub async fn is_mod_or_admin(
   pool: &mut DbPool<'_>,
   local_user_view: &LocalUserView,
   community_id: CommunityId,
 ) -> LemmyResult<()> {
   check_local_user_banned_or_deleted(local_user_view)?;
-  check_is_mod_or_admin(pool, local_user_view.person.id, community_id).await
+  let is_mod = CommunityModeratorView::check_is_community_moderator(
+    pool,
+    community_id,
+    local_user_view.person.id,
+  )
+  .await
+  .is_ok();
+
+  if is_mod || local_user_view.local_user.admin {
+    Ok(())
+  } else {
+    Err(LemmyErrorType::NotAModOrAdmin.into())
+  }
 }
 
 pub async fn is_mod_or_admin_opt(
@@ -148,7 +122,15 @@ pub async fn check_community_mod_of_any_or_admin_action(
   let person = &local_user_view.person;
 
   check_local_user_banned_or_deleted(local_user_view)?;
-  check_is_mod_of_any_or_admin(pool, person.id).await
+  let is_mod_of_any = CommunityModeratorView::is_community_moderator_of_any(pool, person.id)
+    .await
+    .is_ok();
+
+  if is_mod_of_any || local_user_view.local_user.admin {
+    Ok(())
+  } else {
+    Err(LemmyErrorType::NotAModOrAdmin.into())
+  }
 }
 
 pub fn is_admin(local_user_view: &LocalUserView) -> LemmyResult<()> {
